requirements.txt
aiogram==3.2.0
python-dotenv==1.0.0
SQLAlchemy==2.0.23
alembic==1.12.1
aiosqlite==0.19.0
pydantic==2.4.2
python-dateutil==2.8.2
APScheduler==3.10.4 